Tagtype = multiple delete

+++ app/Http/Controllers/TagTypeController.php
@@ -8,6 +8,8 @@ use Yajra\DataTables\Facades\DataTables;
 use Illuminate\Support\Facades\DB;
 use App\Http\Requests\StoreTagTypeRequest;
 use App\Http\Requests\UpdateTagTypeRequest;
+use App\Product;
+use App\Price;
 
 class TagTypeController extends Controller
 {
@@ -116,16 +118,28 @@ class TagTypeController extends Controller
         try {
             DB::beginTransaction();
             
+            // check if tag type has products
+            $productCount = Product::where('tag_type_id', $id)->count();
+            if ($productCount > 0) {
+                return response()->json(['status' => 'error', 'code'=>500,'message' => 'Tag type has products. Please delete products first']);
+            }
+
+            // check if tag type has prices
+            $tagType = Price::where('tag_type_id', $id)->first();
+            if ($tagType) {
+                return response()->json(['status' => 'error', 'code'=>500,'message' => 'Tag type has prices. Please delete prices first']);
+            }
+
             TagType::destroy($id);
 
             DB::commit();
 
-            return response()->json(__('file.The data was successfully deleted'), 200);
+            return response()->json(['status' => 'success', 'code'=>200,'message' => 'Tag type deleted successfully']);
 
         } catch (\Exception $exception) {
             Db::rollBack();
 
-            return response()->json($exception->getMessage(), 500);
+            return response()->json(['status' => 'error', 'code'=>500,'message' => $exception->getMessage()]);
 
         }
         
@@ -135,18 +149,30 @@ class TagTypeController extends Controller
     {
         try {
             DB::beginTransaction();
+
+            // check if tag type has products
+            $productCount = Product::whereIn('tag_type_id', $request->ids)->count();
+            if ($productCount > 0) {
+                return response()->json(['status' => 'error', 'code'=>500,'message' => 'Tag type has products. Please delete products first']);
+            }
+
+            // check if tag type has prices
+            $tagType = Price::whereIn('tag_type_id', $request->ids)->first();
+            if ($tagType) {
+                return response()->json(['status' => 'error', 'code'=>500,'message' => 'Tag type has prices. Please delete prices first']);
+            }
             
             TagType::destroy($request->ids);
 
             DB::commit();
 
-            return response(__('file.Data deleted successfully'), 200);
+            return response()->json(['status' => 'success', 'code'=>200,'message' => 'Tag type deleted successfully']);
 
         } catch (\Exception $exception) {
 
             Db::rollBack();
             
-            return response($exception->getMessage(), 500);
+            return response()->json(['status' => 'error', 'code'=>500,'message' => $exception->getMessage()]);
         }
 
     }
